Cleaned repo of hanging files, adjusted sound volumes, fixed incorrect prompt audio
TODO: I still occasionally suffer from an issue where I immediately trigger the Target Reached condition when entering the second and final round. I think it has something to do with the game engine and how it issues commands and runs them in parallel. I pause the game too, but am still on the hunt for the cause.